+++ lib/elasticity/multi_search_response_parser.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Elasticity
   class MultiSearchResponseParser
     class UnknownError < StandardError; end