+++ lib/elasticity/railtie.rb
@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 require "elasticity/log_subscriber"
 
 module Elasticity
   class Railtie < Rails::Railtie
-    initializer 'elasticity.initialize_logging' do
+    initializer "elasticity.initialize_logging" do
       LogSubscriber.attach_to(:elasticity)
     end
   end