+++ lib/elasticity/strategies.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Elasticity
   module Strategies
     class IndexError < StandardError