+++ spec/units/index_mapper_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "elasticity/index_mapper"
 
 RSpec.describe Elasticity::IndexMapper do