+++ spec/units/multi_search_response_parser_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "elasticity/multi_search_response_parser"
 
 RSpec.describe Elasticity::MultiSearchResponseParser do